requirements.txt
streamlit
Pillow
PyMuPDF
pathlib
numpy 
rembg
opencv-python
python-dotenv
